trellis_m4: Move Pins definition under the 'pins' module
Re-exported from the crate root for compatibility / ease-of-use
